Move handler, state, and onClose callback into custom hook
